refactor: more code smellm
